Refactor to use MenuController instead of AppDelegate
At first I was trying to use the same code in MenuController as in AppDelegate, but that didn't trigger the action/method that I needed when I clicked the NSStatusMenuItem. The fix seems to be assign the item's target to be self.

Also, I factored out showPopover and closePopover methods to clean code up a little.