Add preferences menu
Quit works, setting the frequency doesn't. I'm not quite sure if I made the design proportions correctly for a menu item. The whole thing might be a bit weird with the whole half menu, half field thing. I may scrap it in favor of just popping up a menu or even a fully custom view. Have to get outside opinions.